+++ src/rollup.config.js
@@ -1,3 +1,11 @@
+/**
+ * THIS FILE IS NO LONGER USED. IT IS KEPT FOR REFERENCE ONLY AND WILL BE
+ * DELETED SOON.
+ *
+ * THE EQUIVALENT FILE IS NOW LOCATED AT:
+ * https://github.com/guardian/csnx/tree/main/libs/%40guardian/libs/src/consent-management-platform
+ */
+
 /* eslint-disable import/no-default-export -- it's what rollup wants */
 
 import commonjs from '@rollup/plugin-commonjs';
@@ -6,6 +14,7 @@ import replace from '@rollup/plugin-replace';
 import strip from '@rollup/plugin-strip';
 import typescript from '@rollup/plugin-typescript';
 import css from 'rollup-plugin-css-only';
+import json from '@rollup/plugin-json';
 import pkg from '../package.json';
 
 export default {
@@ -23,12 +32,12 @@ export default {
 	],
 	plugins: [
 		typescript(),
+		json(),
 		css(),
 		resolve(),
 		replace({
 			preventAssignment: true,
 			'process.env.NODE_ENV': JSON.stringify('production'),
-			__PACKAGE_VERSION__: JSON.stringify(pkg.version),
 		}),
 		commonjs(),
 		process.env.NODE_ENV === 'production' &&