+++ src/types/window.d.ts
@@ -1,3 +1,11 @@
+/**
+ * THIS FILE IS NO LONGER USED. IT IS KEPT FOR REFERENCE ONLY AND WILL BE
+ * DELETED SOON.
+ *
+ * THE EQUIVALENT FILE IS NOW LOCATED AT:
+ * https://github.com/guardian/csnx/tree/main/libs/%40guardian/libs/src/consent-management-platform
+ */
+
 import type { getConsentFor } from '../getConsentFor';
 import type { Property } from '../lib/property';
 import type { EndPoint } from '../lib/sourcepointConfig';
@@ -13,16 +21,18 @@ type OnMessageChoiceSelect = (
 	choiceTypeID: 1 | 2 | 3 | 4 | 5 | 6 | 7 | 9 | 10 | 11 | 12 | 13 | 14 | 15,
 ) => void;
 
+type GuCmpHotFix = {
+	initialised?: boolean;
+	cmp?: CMP;
+	onConsent?: typeof onConsent;
+	onConsentChange?: typeof onConsentChange;
+	getConsentFor?: typeof getConsentFor;
+};
+
 declare global {
 	interface Window {
 		// *************** START commercial.dcr.js hotfix ***************
-		guCmpHotFix: {
-			initialised?: boolean;
-			cmp?: CMP;
-			onConsent?: typeof onConsent;
-			onConsentChange?: typeof onConsentChange;
-			getConsentFor?: typeof getConsentFor;
-		};
+		guCmpHotFix: GuCmpHotFix;
 		// *************** END commercial.dcr.js hotfix ***************
 		_sp_queue: [];
 		_sp_?: {
@@ -63,10 +73,7 @@ declare global {
 						message_type: string,
 						pmData: string,
 					) => void;
-					onMessageChoiceError: (
-						message_type: string,
-						err: string,
-					) => void;
+					onMessageChoiceError: (message_type: string, err: string) => void;
 					onPMCancel: (message_type: string) => void;
 					onSPPMObjectReady: () => void;
 					onError: (