+++ CHANGELOG.md
@@ -1,5 +1,54 @@
 # @guardian/consent-management-platform
 
+## 13.12.1
+
+### Patch Changes
+
+- a602daf: Add new comscore cookie to vendorStorageIds
+
+## 13.12.0
+
+### Minor Changes
+
+- da4ad5e: prepare move into csnx
+
+## 13.11.1
+
+### Patch Changes
+
+- bf6d150: Bump package versions as recommended by Dependabot
+
+## 13.11.0
+
+### Minor Changes
+
+- 8b3c987: Adding extra final parameter onConsentChange
+
+## 13.10.0
+
+### Minor Changes
+
+- e506cd9: Adding prebidCustom vendor
+
+## 13.9.0
+
+### Minor Changes
+
+- 0f71443: Removing \_\_gpp stub load in CCPA
+
+### Patch Changes
+
+- 8f5e07c: Bump package version as recommended by Dependabot
+  Update peerDependencies - "@guardian/libs" to "^16.0.0"
+
+## 13.8.0
+
+### Minor Changes
+
+- 60b7d15: - Adding the includeGppApi param to the sourcepoint config for CCPA.
+  - Updating type to include includeGppApi for the window object.
+  - Refactoring stub.ts to load both \_\_uspapi and \_\_gpp stub for CCPA but only load \_\_uspapi for AUS.
+
 ## 13.7.3
 
 ### Patch Changes